move child-process/fs/git/prompt into core/
