+++ core/child-process/README.md
