+++ core/child-process/__tests__/child-process.test.js
