+++ core/child-process/index.js
