+++ core/child-process/package.json
