+++ core/fs-utils/index.js
