+++ core/fs-utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lerna/fs",
+  "name": "@lerna/fs-utils",
   "version": "3.0.0-alpha.3",
   "description": "TODO",
   "keywords": [
