+++ core/git-utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lerna/git",
+  "name": "@lerna/git-utils",
   "version": "3.0.0-alpha.3",
   "description": "TODO",
   "keywords": [
