+++ core/prompt/index.js
