+++ core/prompt/package.json
